Fix module inception to ease merging from main
https://rust-lang.github.io/rust-clippy/master/#/module_inception
