--- indexer/src/storage/mock_db/mock_db.rs
@@ -1,22 +1,0 @@
-use heed::Result;
-
-use crate::storage::types::KVStorageTrait;
-
-pub struct MockDBClient;
-
-#[allow(unused)]
-impl MockDBClient {
-	pub fn new() -> Self {
-		MockDBClient {}
-	}
-}
-
-impl KVStorageTrait for MockDBClient {
-	fn put(&self, _key: &str, _value: &str) -> Result<()> {
-		Ok(())
-	}
-
-	fn get(&self, _key: &str) -> Option<String> {
-		Some("Mock db response".to_string())
-	}
-}
